Правка моделей согласно замечаниям. Удалены лишние базовые модели.
