--- app/Models/Blog/BaseModel.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace App\Models\Blog;
-
-use Illuminate\Database\Eloquent\Model;
-
-/**
- * App\Models\Blog\BaseModel
- *
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\Blog\BaseModel newModelQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\Blog\BaseModel newQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\Blog\BaseModel query()
- * @mixin \Eloquent
- */
-class BaseModel extends Model
-{
-    //
-}
