--- app/Models/Planner/BaseModel.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace App\Models\Planner;
-
-use Illuminate\Database\Eloquent\Model;
-
-/**
- * App\Models\Planner\BaseModel
- *
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\Planner\BaseModel newModelQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\Planner\BaseModel newQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\Planner\BaseModel query()
- * @mixin \Eloquent
- */
-class BaseModel extends Model
-{
-    //
-}
